myProject/requirements.txt
apimatic-core==0.2.15
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
asgiref==3.8.1
CacheControl==0.12.14
certifi==2024.8.30
charset-normalizer==3.4.0
deprecation==2.1.0
Django==5.1.2
django-environ==0.11.2
gunicorn==23.0.0
idna==3.10
jsonpickle==3.0.4
jsonpointer==2.4
msgpack==1.1.0
packaging==24.1
python-dateutil==2.8.2
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
squareup==38.2.0.20241017
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.0
